Improvements for selecting a Rust toolchain
When defining `Rust_TOOLCHAIN`, looking up the
toolchain via `rustup` is prioritized.

When the user changes or sets `Rust_TOOLCHAIN`,
FindRust is re-evaluated.

A non-cache `Rust_TOOLCHAIN` is promoted to a
cache variable.

A toolchain shorthand (e.g. `stable` or `1.44.0` or
`nightly`) can now be specified to Rust_TOOLCHAIN

Options are now documented.

Fixes #37, #43