Add test for cross-compiling cdylib (in particular linking)
Re-use the existing workspace test to add coverage for the scenario of
building a cdylib in Rust and linking it into a CMake executable.

This serves in particular the purpose of adding test-coverage for the
case where cross-compilation of a cdylib needs to pick up the right
linker, so passing the CARGO_TARGET_<triple>_LINKER environment
variable.

In order for this test to work, this also modifies the test-setup
slightly to correctly convey to CMake that we're cross-compiling, by
setting CMAKE_SYSTEM_NAME, which will trigger the setting of
CMAKE_CROSSCOMPILING. And since the C linker is used by default by cargo
for creating cdylibs, we also need to tell CMake about the
cross-compiling C compiler, not just the C++ one.