Fix passing of linker languages
When cross-compiling we always prepend C to the list of linker
languages, unless the target specifies it. A target may also provide
additional linker languages (for the preference matching) through the
DEPS variable. In order to safely pass this list to the generator, it
was originally passed as semi-colon separated list. Commit
546249028d48e7fca03db5fa119c796d0cf85794 broken that by adding
COMMAND_EXPAND_LISTS. This commit fixes that again by changing the
format to be space separated.