Fix unnecessary rebuilds when using CORROSION_USE_HOST_BUILD
Commit 3c9ff60f51cac21fcd4add47c1a14da47fe1c02e introduced the feature
that host binaries can be built, which is done by omitting the
`--target` parameter to cargo. This however implies that the target host
directory is picked by cargo (for example `cargo/build/debug`) for these
builds, which ends up being shared with the corrosion build itself, as
that is built implicitly via `cargo run`. If there were any common
crates as dependencies, then those may end up being built with different
flags (for corrosion and for host builds of the project), resulting in
unnecessary rebuilds.

To fix this, this patch makes use of the internal knowledge about the
host target by always passing it again.