Initial implementation of feature support via target property
This patch builds on the parent commit, but changes the API
to use a target property.

Since this approach requires CMake >= 3.19, the rust2cpp test that received features
was copied into the new features test and reverted from the old test, so that
it can still be run with older cmake versions.