Add --profile support for rust >= 1.57
Add the option to specify the cargo profile by optionally specifying
PROFILE when importing a crate. Cargo profiles are specified on
the workspace level, so it makes sense to select this option with
corrosion_import_crate.
This can be useful for specifying e.g. a `release-with-dbg` or
`release-with-lto` profile etc, that the user can then conditionally
select.

We currently do appear to conditionally link some libraries, depending
on the CMake build mode. So if the user selects a profile that requires
libraries, that are only linked in a different CMake build mode, then
this may cause issues for them. But I would label that a configuration
issue, that can be improved separately if it turns out to be necessary.

Signed-off-by: Jonathan Schwender <jonathan.schwender@huawei.com>