CI: Test on macos-10.15 instead of macos-latest
macos-latest is now macos-11, which is currently not working. The issue
is that `ld` can't find the `System` library.
As a quick fix for the CI it makes sense to just test on the previous
version of macos, until someone finds the time to fix the linker flags
on macos-11

Signed-off-by: Jonathan Schwender <schwenderjonathan@gmail.com>