Remove Integrator build script
Previously the integrator build script was needed to pass the
required libraries and link search paths to cargo.
However, we can do the same by directly specifying the
corresponding rustflags.

This makes life easier for new code, but is a breaking change,
since the build script crate is just removed.
Clients simply have to remove their dependency on the crate,
so the migration is simple enough.