Skip building Generator if possible
Don't build the generator if we are not installing it.

Currently, by default we still install the legacy Generator even if
CMake is >= 3.19, so users can fallback to it in case of any problems
with the new generator. This means that when developing, the generator
will still be built, unless we explicitly opt out.