Downgrade warning to status if no linker pref. is found
There may be no linker preference if the user has no language enabled
in CMake, or no specific compiler selected for a language.
This may be a problem for users who link their library into Rust, but
is not a problem for the other direction.
To remove false warnings downgrade the message to STATUS.
If the wrong linker (language) is selected for Rust code, this should
also cause the link step to fail anyway, and the status should point in the right direction.

Signed-off-by: Jonathan Schwender <schwenderjonathan@gmail.com>