Corrosion tests should not test C++ features not used by Corrosion itself.
This PR follows discussion for #183.

* Problem addressed with this PR:
- CMakeLists.txt uses std_cxx_17 to compile it's test cases.
- Tests cases need std_cxx_17, because of their use of <string_view>.
Because of both, Corrosion fails to compile on 'old' systems (I did not say vintage :-)).

* Investigations:
As Corrosion does not use <string_view>, it's not needed to use this in its test cases.
If <string_view> is not used anywhere, options std_cxx_17 are no more required.

* Correction proposal:
Use of std_cxx_14, instead of std_cxx_17.
Remove the use of <string_view> in Corrosion test cases.
Altogether, this allows older compilers to build Corrosion.

* Remark:
This PR keeps some std_cxx_14 (instead of _17) to keep the need of C++ Standard Library installed on the compilation host.
I did not test, but I suppose std_cxx_11 would work too.