Rename rustflags variable to make global scope clear
The currently available way of setting Rustflags is by setting the `RUSTFLAGS`
environment variable, which affects all crates in the dependency graph.

This may be undesirable in some circumstances, so a future commit will
the option to specify RUSTFLAGS via `cargo rustc`, which only affects
the current crate and not dependencies.

This commit prepares by refactoring some names to make the distinction
easier.