Add support for local Rustflags
For some scenarios it is better to not set Rustflags for all crates
in the dependency graph and instead only set it for the top-level
crate.
For example https://github.com/rust-lang/cargo/issues/8716
can be avoided in some scenarios by setting the rustflags via
rustc, which allows for faster rebuilds in such cases.