Always set IMPORTED_LOCATION
Not setting IMPORTED_LOCATION on imported executables causes CMake
developer warnings. Setting IMPORTED_LOCATION_<CONFIG> is not sufficient.
The issue only affects Multi-Config Generators like Ninja-MultiConfig or
Visual Studio, since otherwise IMPORTED_LOCATION is already always set.

For Multi-Config builds there is no "correct" thing to set
IMPORTED_LOCATION to, so we just always set it, and the last configuration
wins. This is not a problem since the more specific IMPORTED_LOCATION_<CONFIG>
variants are also set in the multi-config case.

This commit refactors a bit of common code into a function, so that both
CorrosionGenerator.cmake and the Rust based CMake Generator can use it.