Backport: Fix finding cargo on NixOS
Backports commit d5330b3f03c7abb4e4da71e35654fa03ecb778bb to stable/v0.2.

cargo on NixOS is not next to `rustc`, so it was previously not found.
To address this, we search for cargo in the regular way via
`find_program`, if the toolchain is not managed by rustup.

Closes #145