Allow explicitly setting the linker for rust executables
Allow explicitly defining a linker for when linking is done on the Rust side.
Previously Corrosion offered an corrosion_set_linker_language(), which
effectively enabled a language, but still determined the linker by choosing
the compiler with the highest priority as determined by CMake.
This is probably not what most users actually want.

Explicitly setting a linker should be less surprising and more in line with
what a user would actually want to do.
Since setting a linker may happen multiple times for different targets and
potentially with different linkers, we don't use find_program to verify
if the linker can actually be found (for performance reasons).

However, we do not allow explicitely setting the linker for MSVC toolchains.
Setting the linker causes strange problems with MSVC, where build-script-build
fails with a file not found os-error, and I'm not exactly sure why. Since with
MSVC the build should anyway be run from a visual studio command prompt, the
correct `cl.exe` will be in PATH, so we can just let cargo find it in this
case.