Fix Rustflags test
Corrosion selects Debug mode if Debug mode or nothing is specified and Release mode
for all other CMake build types.
This commit fixes the rustflags test for other build modes, where
Corrosion selected the release build mode for cargo.
Selecting a custom cargo profile will still cause issues for this test,
but there is not really a reason to do that when building corrosion,
so this can still be improved later.