Revert: Remove --target linker arg
Revert commit cb4e49e240d5c4d58b6117b426264121d8e1deac.
When cross-compiling with clang it is necessary to manually add this
flag. It appears that Rust does not add the flag for us.
CI currently does not test clang + cross-compiling and it's not trivial
to add without doubling the number of tests or reworking the github actions.
Regression testing will be added in a future commit.