Simplify native CMake generator
After the previous refactoring of Corrosion CMake code, we can now reuse
a significant portion of logic and greatly simplify
the Rust application that generates some CMake code.

We still need it for legacy CMake versions to parse JSON, but a lot of
functionality was already implemented in CMake.

This commit removes this code duplication.