Fix MSVC support pre 3.20 with `OUTPUT_DIRECTORY`
The MSVC Generator is Multi-Config and Corrosion offers support for it
even if CMake < 3.20.
The previous commit adding support for respecting `OUTPUT_DIRECTORY`
properties broke this support.

Since special casing MSVC is a lot of effort, add a deprecation notice
that we will require CMake 3.20 in the future for the MSVC generator.
This would allow MSVC to be treated the same as Ninja Multi-Config in
the future and all the special casing could be removed, simplifying
the code.