Add fn to parse package version from Cargo.toml
Add a convenience function to parse the package version from
the package manifest (Cargo.toml). This is mainly intended for
code generators written in Rust, so that downstream users can
easily check if an installed version of the code generator is up
to date with the version specified in the package manifest.
Users could then decide to rebuild from source if the
installed version is outdated and otherwise use the installed version.