Refactor Corrosion tests to use ctest
Use ctest and CMake script mode to drive the corrosion tests.
This has the advantage that we can easily add assertions
on the stdout output of the corrosion configure step,
e.g. causing the test to fail if a Warning occurs when running
Corrosion.
Additionally, we can now also easily run the executables and verify
if they are working, which the tests previously did not check.