FindRust: Remove unused `CARGO_RUST_FLAGS`
These cache variables are currently not used by corrosion
and it is unclear when they were last used.
I think that Corrosion should not silently/magically
add some rustflags based on the CMake profile.
Instead it should be up to the user to define an
appropriate custom cargo profile and map that to
a CMake build type if they wish.