Read OUTPUT_DIRECTORY property from main library target
The target exposed to the users is without the -static or -shared
suffix, so the user is expected to set the OUTPUT_DIRECTORY properties
on that target.

To keep things simple just strip the -static or -shared suffix if
present to determine the target name for looking up the output directory
properties.