Use TARGET signature on cargo-build_<target> to ensure up to date
This is the more conservative solution.
Since  cargo-build_<target> is a custom command and is always out of date, this custom command
will then also be up to date.

If generation time becomes a problem it can be optimized later, and lean on
the safe side for it being up to date for now.

Co-authored-by: Mikhail Svetkin <mikhail.svetkin@remarkable.no>