Fix output_directory support when not used as target property
In CMake, for example RUNTIME_OUTPUT_DIRECTORY is a target property. It
is documented to be initialized from CMAKE_RUNTIME_OUTPUT_DIRECTORY, so
it's not unusual for CMake projects to just set the latter variable in
order to place all affected targets into the directory, instead of
setting a per-target property on each manually.

Since Corrosion respects the target property, it makes sense to assume
that just setting the CMAKE_ variable would also work.