Don't set linker for Rust static libraries
As discussed in issue #270 we at least should not set the linker for
Rust static libraries. In the general case the linker is not needed for
static libraries, and if it is invoked,
Rust should take care of what to choose
(e.g. for building build scripts) instead
of us forcing the choice.