Fix Multi-Config default out dir
Manually appending the configuration type may make
loading of shared libraries fail at runtime.
This was experienced on Linux, where the dynamic loader was
looking for e.g. `Debug/<libname>.so`. The .so was next to the
executable in the Debug directory and could not be found.
To prevent this, we let CMake postfix the Generator
configuration type.