cxxbridge: Relax build dependency
The dependencies of the cxxbridge command should be as follows:
- cxxbridge is available (either installed or built by cmake)
- file level dependency on the rust source file (to regenerate
  if the rust source file changed).
I'm not 100% sure if this is sufficient, potentially we would
also need a dependency on any `include!`ed C header files.

There is no need for any kind of ordering related to the
cargo build command, so remove that dependency.