FindRust: Expose discovered rustup toolchains
The cache entries are only defined if `rustup` is installed.
Users can iterate over the lists using `foreach(... IN ZIP LISTS ..)` to
find a toolchain of interest.
This is useful in the future to let corrosion users override the
toolchain for a specific target by choosing a toolchain from the list
of available toolchains.