Warn when modifying linker options for static libs
If an imported Rust target is only a static lib
it does not make sense to specify a linker or
link libraries into it, because the linker will not be invoked.
Print a warning to inform the user of this likely unwanted behavior.

Since currently both cdylib and staticlib are conflated behind
the same target (and this can't be solved before Rust 1.64) we
omit the warning if a shared library is also built.