Don't force enable a language when cross_compiling
Corrosion itself does not need to enable any languages.
If linking is done via Rust, the rust will by default use
a target specific linker e.g. cc or rust-lld.
If C++ is used then this is not sufficient, but then the
user will also have turned on C++ as a language of their
project.

If linking is handled by CMake, then it is also the
responsibility of the user to enable a language,
otherwise they couldn't compile any code anyway.

We still need to keep the enable_language call in the
main CMakeLists.txt, since FindRust may rely on a compiler
being enabled. FindRust currently unconditionally runs
every time, even though we could just skip through to the
test project (which do enable the required languages).