Remove linker preference
Previoously corrosion set the linker to the C++ compiler in order
to automatically link in the necessary libraries for C++ code.
In practice this seems to be just "c++", so it we can also just
use the default linker and let Corrosion specify "-lc++" as a linker
argument for rust in case the target links in any C++ code.
If the GNU compiler is used for C++ code `libstdc++` is used.

This is generally preferable and should be more in line with users
expect.