Improve Windows target detection
Do not rely on a compiler being enabled.
Instead parse CMAKE_SYSTEM_PROCESSOR, which should fix
wrong detection on windows aarch64. I don't have such a device,
so that is not tested.

We can't rely on a CXX compiler being enabled, so we enhance the
windows ABI check by checking both for known compiler ids
and for the compiler target (e.g. for clang, which supports both).
If no compiler is enabled we fallback to the abi of the default Rust
toolchain.
If we can't infer the ABI, print a warning and fallback to the host
toolchain.

Falling back to the host toolchain is now the default in case something
goes wrong.