Fix imported location target property when target contains dashes and CORROSION_RESPECT_OUTPUT_DIRECTORY is true
The regex to determine the name of the actual library target tries to
strip the -static or -shared suffix did not allow for dashes in the
name. Since the regexp closes with a $, the suffix should only match at
the very end and we can allow any character before the last dash.

Fixes #321