Downgrade log-level for potentially erroneous arguments
Downgrade the log-level to DEBUG for the message informing the
user of missing values for a keyword that requires values.
Passing no values after a keyword may be common in user-code
since it simplifies the usercode.
It may still a sign that something went wrong on the user side,
so we do want to log it. Users investigating problems will then
see the output with the DEBUG log-level enabled.