Improve required library detection for staticlibs
Instead of hand maintaining a list of required libraries, it is more
efficient to just ask `rustc` what libraries are needed
to link.

Using a demo project will not give a full list, but at least a minimum
list of required dependencies - which is the same situation as now.

In the future we could expand this by passing arguments from the cargo
build step to the link-step via a file.
But that needs more investigation, so we do this improvement first.

Expose information about the default host target

Useful when configuring for cross-compilation, but a Rust program must
be compiled for the host target (e.g. code generator).
Having access to information about the host target should be useful for
downstream code (and corrosion itself).