Fix unnecessary recompilations
Changing / unsetting `CARGO_TARGET_xyz_LINKER` causes cargo to
recompile all dependencies, which is unnecessary.
To avoid this we now set the linker directly for the last `rustc`
invocation. This has the side-effect of not affecting the linker
for potential compilations with `cc` in downstream build scripts,
which should be an improvement.

We also now pass `linker-flavor` as an argument, which fixes some
linker inference issues in older rust versions.

Fixes #386