Improve Multi-Config error message
Add the actual CMake version and the path to the used cmake version to
the error message. The CMake version used by the user could be e.g.
a bundled CMake version, which may not be obvious to the user.
Printing the path to the used CMake executable should make such cases
obvious.