Fix: Improve Multi-Config checks
Fixes de3f2f54c7a05984ed1fc7e0cddd91bb07a1b963
by promoting the COR_IS_MULTI_CONFIG variable
to CACHE, so that it stays visible when
`corrosion_import_crate` is called.
It's a bit worrying that this was not caught in CI,
it seems we need some more complex tests.