Fix passing target when linking via clang
When the linker is invoked via Rust, and set to `clang` or a
different compiler that requires a target parameter, the
--target parameter must be passed, otherwise the host linker could
be invoked.

The condition checkingt his was wrong, because CMake evaluates quoted
strings to false, unless the content is explicitely one of the true
values. We need to use the unquoted variable instead in order to have
the desired behavior of checking if the compiler target variable is set.