Refactor CI
Move some of the CI setup into the CMakePresets.json setup.
This should make it more clear that there is no magic going, and
we are just testing many different CMake configurations with corrosion.

Additionally add some needs dependencies to order
some of jobs and skip them if the predecessor failed.