Use cbindgen --depfile feature (#435)
The new v0.25.0 release of cbindgen adds a [`--depfile` option](https://github.com/mozilla/cbindgen/pull/820), that CMake can use to determine when to re-run cbindgen. 
This feature requires the  `DEPFILE` parameter of `add_custom_command` and is available on most generators by now. 
If CMake is too old, or the Generator does not support `DEPFILE`, then we fallback to always rerunning cbindgen as before.