Create CMake targets earlier
Create the main CMake targets earlier, so we can
set target properties on them early in the generator,
so we can use them later on during configuration.

We don't strictly need this right now, but it reduces the diff to
the stable/v0.5 branch and being able to set properties earlier is
something that seems like it will be useful.