Leave choosing `msvcrt` up to the C/C++ CMake code
It's difficult for Corrosion to correctly choose the right
msvcrt version. CMake has a target property for it, and we
also don't know which build configurations the user CMake
code expectes to be Debug related.

The previous code actually already removed the linking
against `msvcrt` in all cases except "Debug" builds.
This was not intentional, but indicates that removing
the library entirely from the link list, and leaving
it up to the C/C++ code to link it in, should work
well in practice.