Leave choosing `msvcrt` up to the C/C++ CMake code (#537)
* Leave choosing `msvcrt` up to the C/C++ CMake code

It's difficult for Corrosion to correctly choose the right
msvcrt version. CMake has a target property for it, and we
also don't know which build configurations the user CMake
code expectes to be Debug related.

The previous code actually already removed the linking
against `msvcrt` in all cases except "Debug" builds.
This was not intentional, but indicates that removing
the library entirely from the link list, and leaving
it up to the C/C++ code to link it in, should work
well in practice.

Co-authored-by: kotoriのねこ <minamiktr@outlook.com>
(cherry picked from commit b45275c8b5ee16118008a6a140ce842557b3e47a)