Propagate Apple deployment target version from CMake to `cc`
`rustc` always targets the earliest deployment platform version possible
for Rust code, but the `cc` crate defaults to the host's SDK version
which is typically very recent.

When the CMake uses a toolchain that has its' own minimal supported macOS
version (for example Qt), this leads to linker errors and possible crashes
when deploying to older yet supported systems.