Guard CORROSION_TOOLS_RUST_TOOLCHAIN behind rustup check
If the toolchain is not managed by rustup,
then resolving the toolchain would fail,
hence we instead ignore the option
and fallback to the default toolchain.
Also simplify the check for cargo / rustc.
We check once in FindRust, then we
can later just check if the variable evaulates to true or not.