cbindgen: Fix adding header dependency on cargo-build
If headers are generated with cbindgen, they should
probably always be available if the rust crate is also
available, so it makes sense to add a dependency.
The changed line was supposed to already do exactly that,
however interface targets are not actually built,
so we need to directly depend on the cargo-build_
custom command that generates the Rust library.
We already do this in other places, it was just
overlooked here.